Working baddies bouncing off each other
I solved this by putting the map of baddies in a bank and then using
PEEK & POKE to figure out who's where.

The code has a lot of replication, but when it was calling a DEFPROCs
many times, the game play was too slow.

So instead I inline the code to each function. Feels wasteful, but works